+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### 3.9.0
+  * Add pagination support for fulfillment line items [#248](https://github.com/singer-io/tap-shopify/pull/248)
+
+### 3.8.1
+  * Update the bookmark value even if the record is not retrieved [#247](https://github.com/singer-io/tap-shopify/pull/247)
+
+### 3.8.0
+  * Exponential backoff for Shopify bulk operations in progress [#246](https://github.com/singer-io/tap-shopify/pull/246)
+
 ### 3.7.3
   * Cleanup state file for failed Bulk operations [#244](https://github.com/singer-io/tap-shopify/pull/244)